requirements.txt
numpy==1.19.1
python-Levenshtein==0.12.0
scikit-learn==0.23.1
scipy==1.5.0
tensorboard==1.13.1
tensorflow-gpu==1.13.1